getAreaDrops를 areaController로 옮김
